fix: image loading. optimised overside image.
