+++ src/pages/partners/affiliate-ib/_hero.tsx
@@ -17,9 +17,6 @@ const query = graphql`
         image: file(relativePath: { eq: "affiliate/partners-banner.png" }) {
             ...heroImage
         }
-        image_mobile: file(relativePath: { eq: "affiliate/partners-banner-mobile.png" }) {
-            ...heroImage
-        }
     }
 `
